move purchase content form state to common
