+++ packages/common/src/hooks/index.ts
@@ -20,3 +20,4 @@ export * from './useSavedCollections'
 export * from './chats'
 export * from './useGeneratePlaylistArtwork'
 export * from './useUSDCBalance'
+export * from './purchaseContent'
