+++ packages/common/src/hooks/purchaseContent/utils.ts
@@ -1,0 +1,22 @@
+import { payExtraAmountPresetValues } from './constants'
+import { PayExtraPreset } from './types'
+
+export const getExtraAmount = (
+  amountPreset: PayExtraPreset,
+  customAmount = 0
+) => {
+  let extraAmount = 0
+  switch (amountPreset) {
+    case PayExtraPreset.LOW:
+    case PayExtraPreset.MEDIUM:
+    case PayExtraPreset.HIGH:
+      extraAmount = payExtraAmountPresetValues[amountPreset]
+      break
+    case PayExtraPreset.CUSTOM:
+      extraAmount = Number.isFinite(customAmount) ? customAmount : 0
+      break
+    default:
+      break
+  }
+  return extraAmount
+}
