+++ packages/common/src/hooks/purchaseContent/validation.ts
