+++ packages/mobile/src/components/premium-track-purchase-drawer/PremiumTrackPurchaseDrawer.tsx
@@ -10,14 +10,17 @@ import {
   combineStatuses,
   useUSDCBalance,
   getPurchaseSummaryValues,
-  statusIsNotFinalized
+  statusIsNotFinalized,
+  isContentPurchaseInProgress,
+  ContentType
 } from '@audius/common'
+import { Formik } from 'formik'
 import { Linking, View } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 
 import IconCart from 'app/assets/images/iconCart.svg'
 import IconError from 'app/assets/images/iconError.svg'
-import { LockedStatusBadge, Text } from 'app/components/core'
+import { Button, LockedStatusBadge, Text } from 'app/components/core'
 import { NativeDrawer } from 'app/components/drawer'
 import { useDrawer } from 'app/hooks/useDrawer'
 import { useIsUSDCEnabled } from 'app/hooks/useIsUSDCEnabled'
@@ -30,14 +33,16 @@ import { TrackDetailsTile } from '../track-details-tile'
 
 import { PurchaseSuccess } from './PurchaseSuccess'
 import { PurchaseSummaryTable } from './PurchaseSummaryTable'
-import { StripePurchaseConfirmationButton } from './StripePurchaseConfirmationButton'
 
 const { getPurchaseContentError, getPurchaseContentFlowStage } =
   purchaseContentSelectors
 
+const { startPurchaseContentFlow } = purchaseContentActions
+
 const PREMIUM_TRACK_PURCHASE_MODAL_NAME = 'PremiumTrackPurchase'
 
 const messages = {
+  buy: (price: string) => `Buy $${price}`,
   title: 'Complete Purchase',
   summary: 'Summary',
   artistCut: 'Artist Cut',
@@ -47,6 +52,7 @@ const messages = {
   youPaid: 'You Paid',
   price: (price: string) => `$${price}`,
   payToUnlock: 'Pay-To-Unlock',
+  purchasing: 'Purchasing',
   disclaimer: (termsOfUse: ReactNode) => (
     <>
       {'By clicking on "Buy", you agree to our '}
@@ -112,7 +118,8 @@ const useStyles = makeStyles(({ spacing, typography, palette }) => ({
 
 export const PremiumTrackPurchaseDrawer = () => {
   const styles = useStyles()
-  const { neutralLight2, accentRed, secondary } = useThemeColors()
+  const { specialLightGreen, neutralLight2, accentRed, secondary } =
+    useThemeColors()
   const dispatch = useDispatch()
   const isUSDCEnabled = useIsUSDCEnabled()
   const { data } = useDrawer('PremiumTrackPurchase')
@@ -124,11 +131,21 @@ export const PremiumTrackPurchaseDrawer = () => {
   const { data: currentBalance, status: balanceStatus } = useUSDCBalance()
   const error = useSelector(getPurchaseContentError)
   const stage = useSelector(getPurchaseContentFlowStage)
+  const isPurchasing = isContentPurchaseInProgress(stage)
   const isPurchaseSuccessful = stage === PurchaseContentStage.FINISH
   const { premium_conditions: premiumConditions } = track ?? {}
   const isLoading = statusIsNotFinalized(
     combineStatuses([trackStatus, balanceStatus])
   )
+
+  const handleConfirmPurchase = useCallback(() => {
+    dispatch(
+      startPurchaseContentFlow({
+        contentId: trackId,
+        contentType: ContentType.TRACK
+      })
+    )
+  }, [dispatch, trackId])
 
   const handleClosed = useCallback(() => {
     dispatch(purchaseContentActions.cleanup())
@@ -161,58 +178,72 @@ export const PremiumTrackPurchaseDrawer = () => {
           <LoadingSpinner />
         </View>
       ) : (
-        <View style={styles.drawer}>
-          <View style={styles.titleContainer}>
-            <IconCart fill={neutralLight2} />
-            <Text style={styles.title}>{messages.title}</Text>
-          </View>
-          <TrackDetailsTile trackId={track.track_id} />
-          <PurchaseSummaryTable
-            {...purchaseSummaryValues}
-            isPurchaseSuccessful={isPurchaseSuccessful}
-          />
-          {isPurchaseSuccessful ? (
-            <PurchaseSuccess track={track} />
-          ) : (
-            <>
-              <View>
-                <View style={styles.payToUnlockTitleContainer}>
-                  <Text
-                    weight='heavy'
-                    textTransform='uppercase'
-                    fontSize='small'
-                  >
-                    {messages.payToUnlock}
-                  </Text>
-                  <LockedStatusBadge locked />
-                </View>
-                <Text style={styles.disclaimer}>
-                  {messages.disclaimer(
-                    <Text colorValue={secondary} onPress={handleTermsPress}>
-                      {messages.termsOfUse}
-                    </Text>
-                  )}
-                </Text>
+        <Formik onSubmit={handleConfirmPurchase}>
+          {({ values, handleSubmit }) => (
+            <View style={styles.drawer}>
+              <View style={styles.titleContainer}>
+                <IconCart fill={neutralLight2} />
+                <Text style={styles.title}>{messages.title}</Text>
               </View>
-              <StripePurchaseConfirmationButton
-                trackId={track.track_id}
-                price={formatPrice(price)}
+              <TrackDetailsTile trackId={track.track_id} />
+              <PurchaseSummaryTable
+                {...purchaseSummaryValues}
+                isPurchaseSuccessful={isPurchaseSuccessful}
               />
-            </>
-          )}
-          {error ? (
-            <View style={styles.errorContainer}>
-              <IconError
-                fill={accentRed}
-                width={spacing(5)}
-                height={spacing(5)}
-              />
-              <Text weight='medium' colorValue={accentRed}>
-                {messages.error}
-              </Text>
+              {isPurchaseSuccessful ? (
+                <PurchaseSuccess track={track} />
+              ) : (
+                <>
+                  <View>
+                    <View style={styles.payToUnlockTitleContainer}>
+                      <Text
+                        weight='heavy'
+                        textTransform='uppercase'
+                        fontSize='small'
+                      >
+                        {messages.payToUnlock}
+                      </Text>
+                      <LockedStatusBadge locked />
+                    </View>
+                    <Text style={styles.disclaimer}>
+                      {messages.disclaimer(
+                        <Text colorValue={secondary} onPress={handleTermsPress}>
+                          {messages.termsOfUse}
+                        </Text>
+                      )}
+                    </Text>
+                  </View>
+                  <Button
+                    disabled={isPurchasing}
+                    title={
+                      isLoading
+                        ? messages.purchasing
+                        : messages.buy(formatPrice(price))
+                    }
+                    variant={'primary'}
+                    size='large'
+                    color={specialLightGreen}
+                    iconPosition='left'
+                    icon={isLoading ? LoadingSpinner : undefined}
+                    fullWidth
+                  />
+                </>
+              )}
+              {error ? (
+                <View style={styles.errorContainer}>
+                  <IconError
+                    fill={accentRed}
+                    width={spacing(5)}
+                    height={spacing(5)}
+                  />
+                  <Text weight='medium' colorValue={accentRed}>
+                    {messages.error}
+                  </Text>
+                </View>
+              ) : null}
             </View>
-          ) : null}
-        </View>
+          )}
+        </Formik>
       )}
     </NativeDrawer>
   )
